Optimización al codigó

+++ src/04_Zero-Values/main.go
@@ -4,7 +4,7 @@ import "fmt"
 
 func main() {
 	// Zero values
-	var a int     // Default = 0
+	var a uint8   // Default = 0
 	var b float64 // Default = 0
 	var c string  // String Vacio
 	var d bool    // Default = False
